refactor(server): share toPublicNode for the upcoming page-model read path
